requirements.txt
requests==2.31.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pytest==7.4.0
flake8==6.1.0
black==23.7.0
mypy==1.5.1